Semen

+++ src/services/userServices/MentorService.ts
@@ -1,10 +1,12 @@
 import { Op } from "sequelize";
+import { deleteMentorPictureByName } from "../../lib/utils/utils";
 import { Mentors } from "../../models/models";
 
 interface IMentor {
     mentor_fullName: string, 
     mentor_info: string, 
-    mentor_email: string
+    mentor_email: string,
+    mentor_picture: string
 }
 
 export default class MentorService {
@@ -37,10 +39,14 @@ export default class MentorService {
     }
 
     async deleteMentor(id: number) {
-        Mentors.destroy({
-            where: {
-                id_mentor: id
-            }
-        })
+        let mentor = await this.getMentorById(id)
+        if (mentor) {
+            deleteMentorPictureByName(mentor.get('mentor_picture') as string)
+            Mentors.destroy({
+                where: {
+                    id_mentor: id
+                }
+            })
+        }
     }
 }
